Remove redundant cast in AotMethods

+++ aot-runtime/src/main/java/com/dylibso/chicory/aot/runtime/AotMethods.java
@@ -31,7 +31,7 @@ public final class AotMethods {
     }
 
     public static int tableGet(int index, int tableIndex, Instance instance) {
-        return (int) OpcodeImpl.TABLE_GET(instance, tableIndex, index);
+        return OpcodeImpl.TABLE_GET(instance, tableIndex, index);
     }
 
     public static void tableSet(int index, int value, int tableIndex, Instance instance) {
